Move contiguous split declarations to contiguous_split.hpp
