+++ cpp/src/copying/pack.cpp
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-#include <cudf/detail/copy.hpp>
+#include <cudf/detail/contiguous_split.hpp>
 #include <cudf/detail/nvtx/ranges.hpp>
+#include <cudf/contiguous_split.hpp>
 #include <cudf/utilities/default_stream.hpp>
 
 #include <rmm/cuda_stream_view.hpp>
