+++ python/cudf/cudf/_lib/cpp/copying.pxd
@@ -18,12 +18,6 @@ from cudf._lib.exception_handler cimport cudf_exception_handler
 
 ctypedef const scalar constscalar
 
-cdef extern from "cudf/copying.hpp" namespace "cudf::packed_columns" nogil:
-    cdef struct metadata:
-        metadata(vector[uint8_t]&& v)
-        const uint8_t* data () except +
-        size_type size () except +
-
 cdef extern from "cudf/copying.hpp" namespace "cudf" nogil:
     ctypedef enum out_of_bounds_policy:
         NULLIFY 'cudf::out_of_bounds_policy::NULLIFY'
@@ -112,23 +106,6 @@ cdef extern from "cudf/copying.hpp" namespace "cudf" nogil:
         table_view input_table,
         vector[size_type] splits
     ) except +
-
-    cdef cppclass packed_columns:
-        unique_ptr[metadata] metadata_
-        unique_ptr[device_buffer] gpu_data
-
-    cdef struct contiguous_split_result:
-        table_view table
-        vector[device_buffer] all_data
-
-    cdef vector[contiguous_split_result] contiguous_split (
-        table_view input_table,
-        vector[size_type] splits
-    ) except +
-
-    cdef packed_columns pack (const table_view& input) except +
-
-    cdef table_view unpack (const packed_columns& input) except +
 
     cdef unique_ptr[column] copy_if_else (
         column_view lhs,
